save each class in seperate file
