+++ teacher.rb
@@ -1,0 +1,12 @@
+require_relative 'person'
+
+class Teacher < Person
+    def initialize(id, specialization, age = nil, name: "Unknown", parent_permission: true)
+      super(id, age, name: name, parent_permission: parent_permission)
+      @specialization = specialization
+    end
+  
+    def can_use_services?
+      true
+    end
+  end
